restructure view into a new file
